chore(console): move custom domain schema definition
